% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encoding.R
\name{assign.a}
\alias{assign.a}
\title{For affecteds: Take genetic variant and determine the category of the combo.}
\usage{
assign.a(variant)
}
\arguments{
\item{variant}{Variant for individual. genotypes, phased genotypes, or binary encodings accepted.}
}
\description{
For affecteds: Take genetic variant and determine the category of the combo.
}
