% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encoding.R
\name{assign.u}
\alias{assign.u}
\title{For unaffecteds:  Take a genetic variant and determine the category of the combo.}
\usage{
assign.u(variant)
}
\arguments{
\item{variant}{Variant for individual. genotypes, phased genotypes, or binary encodings accepted.}
}
\description{
For unaffecteds:  Take a genetic variant and determine the category of the combo.
}
