% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encoding.R
\name{encode.rows}
\alias{encode.rows}
\title{Take the relationship matrix and the encoded statuses of info.
For each row, generate the encoded data for scoring.}
\usage{
encode.rows(relation.mat, status.df, ...)
}
\description{
Take the relationship matrix and the encoded statuses of info.
For each row, generate the encoded data for scoring.
}
