% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree.r
\name{read.pedigree}
\alias{read.pedigree}
\title{Read in the encoded pedigree data file.
TODO - can we build this encoded matrix from a more standard file format? i.e. ped-like?}
\usage{
read.pedigree(filename)
}
\arguments{
\item{filename}{name of the file with the data.}
}
\value{
A data frame containing the encoded pedigree information.
}
\description{
Read in the encoded pedigree data file.
TODO - can we build this encoded matrix from a more standard file format? i.e. ped-like?
}
\examples{
example.pedigree.file <-system.file('extdata/example_pedigree_encoding.tsv', package = 'seqbio.variant.scoring')
example.pedigree.df <- read.pedigree(example.pedigree.file)
}
