% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read.relation.mat}
\alias{read.relation.mat}
\title{Read in relationship matrix
Apply the individual names to the rows and columns.}
\usage{
read.relation.mat(fname, indiv.df)
}
\description{
Row/column intersections give the degree of relationship for the
two individuals. 0 = self, -1 = unrelated.
}
