% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree.r
\name{score}
\alias{score}
\title{Score the pedigrees using the pihat values.}
\usage{
score(pihat, K = -1)
}
\arguments{
\item{pihat}{Estimated proportion of genome shared between individuals, from function: ibd.}

\item{K}{Estimated penetrance value, from function: penetrance.}
}
\description{
Score the pedigrees using the pihat values.
}
