% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relatedness.r
\name{sum.fam.scores}
\alias{sum.fam.scores}
\title{Sum all the given scores and return a single vector with cumulative "score", "for" and "against" vals.
For use in instances where one wishes to combine scores from multiple families.}
\usage{
\method{sum}{fam.scores}(score.vec)
}
\description{
Sum all the given scores and return a single vector with cumulative "score", "for" and "against" vals.
For use in instances where one wishes to combine scores from multiple families.
}
