% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encoding.R
\name{assign.status}
\alias{assign.status}
\title{Take a disease status and a genetic variant and determine which category the combo falls in.
A_c = Affected individual with ALT variant
A_i = Affected individual without ALT variant
U_c = Unaffected individual without ALT variant
U_i = Unaffected individualwith ALT variant
If theoretical.max = TRUE the true variant statuses are ignored and all
affected/unaffected are assigned A_c and U_c respectively.
These encodings can then be used show what a family's max score would be.}
\usage{
assign.status(status, variant, theoretical.max = FALSE)
}
\description{
Take a disease status and a genetic variant and determine which category the combo falls in.
A_c = Affected individual with ALT variant
A_i = Affected individual without ALT variant
U_c = Unaffected individual without ALT variant
U_i = Unaffected individualwith ALT variant
If theoretical.max = TRUE the true variant statuses are ignored and all
affected/unaffected are assigned A_c and U_c respectively.
These encodings can then be used show what a family's max score would be.
}
