% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encoding.R
\name{build.relation.dict}
\alias{build.relation.dict}
\title{Build dictonary with the relationships falling in the different categories for the query row.}
\usage{
build.relation.dict(mat.row, name.stat.dict, drop.unrelated = TRUE)
}
\description{
Build dictonary with the relationships falling in the different categories for the query row.
}
