% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relatedness.r
\name{calc.rv.score}
\alias{calc.rv.score}
\title{Calculate a relatedness-weighted score for a given rare variant.}
\usage{
calc.rv.score(
  fam_list,
  affected.weight = 1,
  unaffected.weight = 0.5,
  unaffected_max = 8
)
}
\description{
These scores can be used to compare variants of interest within a family
}
\details{
For each individual, a relationship-informed weight is applied to their sharing
or not sharing of a variant.
The score for affected status is:
(1 / coefficient_of_relatedness) * status_weight
For example, an affected cousin (encoded as a 3) would get a score of:
(1/0.125) * affected weight
8 * 1
= 8 points in favour of the variant.
Whereas for unaffected unaffected individuals, scores decay the further a person is in
relation to the proband based on the formula:
((unaffected_max\emph{2) * coefficient_of_relatedness ) * unaffected_weight
For example, with the default unaffected_max of 8. The sister that does not have a variant would get a score of
((8}2) 0.5) * unaffected_weight
(16 * 0.5) * 0.5
= 4 points for the variant.
If these were the only two relatives considered we could sum the points
and get a score in favour of the variant of
8 + 4 = 12
Ig there is evidence against a variant, this is factored into the score as:
total_score = evidence_for - evidence_against
For example, if there were also an affected sibling without the variant we would have the score against of:
(1/0.5) * 1 = 2
The final score for the variant would then be
.    for - against = total
12 - 2 = 10
Giving a final score of 10 for the variant. Comparing values across variants can be used
to rank them based on pedigree-informed levels of variant sharing across affected
and unaffected individuals.

Input:
fam_dict
- A dictionary with the keys: \link{'A_c', 'A_i', 'U_c', 'U_i'}
respectively containing the affected correct, affected incorrect,
unaffected correct and unaffected incorrect.
- where each value in the dictionary is a list containing the proband and the
proband's relatives as encoded based on their degree of relatedness to the
proband. (proband = 0, sibling/parent/offspring = 1, uncle/grandparent = 2,
cousin = 3, etc.)
affected.weight
- a coefficient to multiply the calculated A_c and A_i relatedness values by.
unaffected.weight
- a coefficient to multiply the U_c and U_i relatedness values by.
unaffected_max
- is the param controlling the score given to a first degree unaffected relatives
scores decay from this specified maximum by half for each subsequent relationship degree.

\if{html}{\out{<div class="sourceCode">}}\preformatted{Increasing the affected.weight relative to the unaffected.weight will make the scores
give more weight to the correct/incorrect status of affected individuals. The default
is 2:1 weight for affected relative to unaffected, which accounts for the fact that
variants are likely to be incompletely penetrant and we therefore want to be more tolerant
of unaffected individuals that have a variant rather than affected individuals that do not.
}\if{html}{\out{</div>}}
}
