% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encoding.R
\name{encode.rows}
\alias{encode.rows}
\title{Take the relationship matrix and the encoded statuses of info.
For each row, generate the encoded data for scoring.}
\usage{
encode.rows(relation.mat, status.df, ...)
}
\arguments{
\item{relation.mat}{The relationship matrix for all pairwise combinations of individuals.}

\item{status.df}{The ID, status, and genotypes for each individual.}
}
\value{
A dictionary with the per-individual relationship lists.
One value for each row of the matrix.
}
\description{
Take the relationship matrix and the encoded statuses of info.
For each row, generate the encoded data for scoring.
}
\examples{
#TODO - add
}
