% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read.indiv}
\alias{read.indiv}
\title{Read in variant and status info for individuals.}
\usage{
read.indiv(fname)
}
\arguments{
\item{fname}{A file name, expected format of contents is:
name             status    variant
MS-4107-1001      A      0/1}
}
\value{
A dataframe.
}
\description{
Read in variant and status info for individuals.
}
\examples{
tsv.name1 <-system.file('extdata/7003_notch3.tsv', package = 'seqbio.variant.scoring')
id.df1 <- read.indiv(tsv.name1)
}
