% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree.r
\name{read.pedigree}
\alias{read.pedigree}
\title{Read in the encoded pedigree data file.}
\usage{
read.pedigree(filename)
}
\arguments{
\item{filename}{name of the file with the data.}
}
\value{
A data frame containing the encoded pedigree information.
}
\description{
Read in the encoded pedigree data file.
}
\examples{
example.pedigree.file <- system.file('extdata/example_pedigree_encoding.tsv',
package = 'KinformR')
example.pedigree.df <- read.pedigree(example.pedigree.file)
}
