% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read.relation.mat}
\alias{read.relation.mat}
\title{Read in relationship matrix
Apply the individual names to the rows and columns.}
\usage{
read.relation.mat(fname)
}
\arguments{
\item{fname}{The file with the relationship matrix information.}
}
\value{
A matrix with the relationships and individual ids as rownames and colnames.
}
\description{
Row/column intersections give the degree of relationship for the
two individuals. 0 = self, -1 = unrelated.
}
\examples{
mat.name1 <-system.file('extdata/7003_notch3.mat', package = 'seqbio.variant.scoring')
mat1 <- read.relation.mat(mat.name1)
}
