% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree.r
\name{score}
\alias{score}
\title{Score the pedigrees using the pihat values.}
\usage{
score(pihat)
}
\arguments{
\item{pihat}{Estimated proportion of genome shared between individuals,
from function: ibd.}
}
\value{
The score value.
}
\description{
Score the pedigrees using the pihat values.
}
\examples{
s.val <- score(12.61)
}
