% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relatedness.r
\name{subset.mat}
\alias{subset.mat}
\title{Take the matrix and subset out only the encoded individuals that are present in the status dataframe.}
\usage{
\method{subset}{mat}(mat.df, status.df)
}
\description{
Take the matrix and subset out only the encoded individuals that are present in the status dataframe.
}
